requirements.txt
streamlit
google-generativeai
python-dotenv
fpdf
matplotlib
Pillow
